refactor: rename components and separate them into base and example components.
