+++ docs/.vuepress/baseComponents/IconSizesTable.vue
