+++ docs/.vuepress/baseComponents/WeatherCodesTable.vue
