+++ docs/.vuepress/clientAppEnhance.js
@@ -1,16 +1,19 @@
 import {defineClientAppEnhance} from '@vuepress/client';
 
-import CodeWellHeader from './components/CodeWellHeader.vue';
+// Common views
 import Icons from "./views/Icons.vue";
 import Colors from "./views/Colors.vue";
-import UtilityClassTable from "./components/UtilityClassTable.vue";
-import ComponentClassTable from "./components/ComponentClassTable.vue";
-import WeatherCodesTable from "./components/WeatherCodesTable.vue";
-import SpacingSystemTable from "./components/SpacingSystemTable.vue";
-import IconSizesTable from "./components/IconSizesTable.vue";
-import ComponentAccessibleTable from "./components/ComponentAccessibleTable.vue";
 
-// ICons
+// Base components
+import CodeWellHeader from './baseComponents/CodeWellHeader.vue';
+import UtilityClassTable from "./baseComponents/UtilityClassTable.vue";
+import ComponentClassTable from "./baseComponents/ComponentClassTable.vue";
+import WeatherCodesTable from "./baseComponents/WeatherCodesTable.vue";
+import SpacingSystemTable from "./baseComponents/SpacingSystemTable.vue";
+import IconSizesTable from "./baseComponents/IconSizesTable.vue";
+import ComponentAccessibleTable from "./baseComponents/ComponentAccessibleTable.vue";
+
+// Common icons
 import IconInfo from '@svgIcons/IconInfo.vue';
 import IconLightbulb from "@svgIcons/IconLightbulb.vue";
 import IconCheckCircle from "@svgIcons/IconCheckCircle.vue";
@@ -19,21 +22,27 @@ import IconError from "@svgIcons/IconError.vue";
 import IconClose from "@svgIcons/IconClose.vue";
 import IconCheckmark from "@svgIcons/IconCheckmark.vue";
 
+// Less files
 import '../.././lib/build/less/dialtone.less';
 import './styles/dialtone-docs.less';
 import './styles/overrides.less';
 import './styles/hljs-dialpad.less';
 
 export default defineClientAppEnhance(({app, router, siteData}) => {
+    // Common views
+    app.component('icons', Icons);
+    app.component('colors', Colors);
+
+    // Base components
+    app.component('code-well-header', CodeWellHeader);
+    app.component('utility-class-table', UtilityClassTable);
     app.component('component-class-table', ComponentClassTable);
     app.component('weather-codes-table', WeatherCodesTable);
-    app.component('utility-class-table', UtilityClassTable);
     app.component('spacing-system-table', SpacingSystemTable);
     app.component('icon-sizes-table', IconSizesTable);
     app.component('component-accessible-table', ComponentAccessibleTable);
-    app.component('code-well-header', CodeWellHeader);
-    app.component('icons', Icons);
-    app.component('colors', Colors);
+
+    // Common icons
     app.component('icon-info', IconInfo);
     app.component('icon-lightbulb', IconLightbulb);
     app.component('icon-check-circle', IconCheckCircle);
